[TESTING] use Float for CGFloat.NativeType

+++ Foundation/CGFloat.swift
@@ -16,7 +16,7 @@ public struct CGFloat {
 #elseif arch(x86_64) || arch(arm64) || arch(s390x)
     /// The native type used to store the CGFloat, which is Float on
     /// 32-bit architectures and Double on 64-bit architectures.
-    public typealias NativeType = Double
+    public typealias NativeType = Float
 #endif
     
     @_transparent public init() {
@@ -169,7 +169,7 @@ public struct CGFloat {
 #if arch(i386) || arch(arm)
         native = NativeType(bitPattern: UInt32(bitPattern))
 #elseif arch(x86_64) || arch(arm64) || arch(s390x)
-        native = NativeType(bitPattern: UInt64(bitPattern))
+        native = NativeType(bitPattern: UInt32(bitPattern))
 #endif
     }
